S09 p22 b208 274 바코드 고유 번호 생성 api 구현

See merge request s09-fintech-finance-sub2/S09P22B208!97

+++ back/src/main/java/com/cocopay/payment/dto/res/CardOfferResDto.java
@@ -42,7 +42,7 @@ public class CardOfferResDto {
     private int remainingAmt;
 
     //실적 그래프 퍼센트
-    private int graphRate;
+    private double graphRate;
 
     //전월실적 달성 여부 -> 내가 이번 달 혜택을 받을 수 있는지
     private boolean pastPerformance;
